# proto/api/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/api/api.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/getActions": {
      "post": {
        "summary": "get action(s) by:\n1. start index and action count\n2. action hash\n3. address with start index and action count\n4. get unconfirmed actions by address with start index and action count\n5. block hash with start index and action count",
        "operationId": "GetActions",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/iotexapiGetActionsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/iotexapiGetActionsRequest"
            }
          }
        ],
        "tags": [
          "APIService"
        ]
      }
    },
...
    "/v1/streamBlocks": {
      "post": {
        "summary": "get block info in stream",
        "operationId": "StreamBlocks",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/iotexapiStreamBlocksResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/iotexapiStreamBlocksRequest"
            }
          }
        ],
        "tags": [
          "APIService"
        ]
      }
    },
    "/v1/streamLogs": {
      "post": {
        "summary": "get logs filtered by contract address and topics in stream",
        "operationId": "StreamLogs",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "$ref": "#/x-stream-definitions/iotexapiStreamLogsResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/iotexapiStreamLogsRequest"
            }
          }
        ],
        "tags": [
          "APIService"
        ]
      }
    }
  },
  "definitions": {
    "iotexapiActionInfo": {
      "type": "object",
      "properties": {
        "action": {
          "$ref": "#/definitions/iotextypesAction"
        },
        "actHash": {
          "type": "string"
        },
        "blkHash": {
          "type": "string"
        },
        "blkHeight": {
          "type": "string",
          "format": "uint64"
        },
        "sender": {
          "type": "string"
        },
        "gasFee": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "iotexapiGetActionByHashRequest": {
      "type": "object",
      "properties": {
        "actionHash": {
          "type": "string"
        },
        "checkPending": {
          "type": "boolean",
          "format": "boolean"
        }
      }
    },
    "iotexapiGetActionsByAddressRequest": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "start": {
          "type": "string",
          "format": "uint64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotexapiGetActionsByBlockRequest": {
      "type": "object",
      "properties": {
        "blkHash": {
          "type": "string"
        },
        "start": {
          "type": "string",
          "format": "uint64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotexapiGetActionsByIndexRequest": {
      "type": "object",
      "properties": {
        "start": {
          "type": "string",
          "format": "uint64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotexapiGetActionsRequest": {
      "type": "object",
      "properties": {
        "byIndex": {
          "$ref": "#/definitions/iotexapiGetActionsByIndexRequest"
        },
        "byHash": {
          "$ref": "#/definitions/iotexapiGetActionByHashRequest"
        },
        "byAddr": {
          "$ref": "#/definitions/iotexapiGetActionsByAddressRequest"
        },
        "unconfirmedByAddr": {
          "$ref": "#/definitions/iotexapiGetUnconfirmedActionsByAddressRequest"
        },
        "byBlk": {
          "$ref": "#/definitions/iotexapiGetActionsByBlockRequest"
        }
      }
    },
    "iotexapiGetActionsResponse": {
      "type": "object",
      "properties": {
        "total": {
          "type": "string",
          "format": "uint64"
        },
        "actionInfo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/iotexapiActionInfo"
          }
        }
      }
    },
...
    "iotexapiGetUnconfirmedActionsByAddressRequest": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "start": {
          "type": "string",
          "format": "uint64"
        },
        "count": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "iotexapiLogsFilter": {
      "type": "object",
      "properties": {
        "address": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "topics": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/iotexapiTopics"
          }
        }
      }
    },
...
    "iotexapiStreamBlocksRequest": {
      "type": "object",
      "title": "below are streaming APIs"
    },
...
    "iotexapiStreamLogsRequest": {
      "type": "object",
      "properties": {
        "filter": {
          "$ref": "#/definitions/iotexapiLogsFilter"
        }
      }
    },
...
    "iotexapiTopics": {
      "type": "object",
      "properties": {
        "topic": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    },
...
    "iotextypesAction": {
      "type": "object",
      "properties": {
        "core": {
          "$ref": "#/definitions/iotextypesActionCore"
        },
        "senderPubKey": {
          "type": "string",
          "format": "byte"
        },
        "signature": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "iotextypesActionCore": {
      "type": "object",
      "properties": {
        "version": {
          "type": "integer",
          "format": "int64"
        },
        "nonce": {
          "type": "string",
          "format": "uint64"
        },
        "gasLimit": {
          "type": "string",
          "format": "uint64"
        },
        "gasPrice": {
          "type": "string"
        },
        "transfer": {
          "$ref": "#/definitions/iotextypesTransfer"
        },
        "execution": {
          "$ref": "#/definitions/iotextypesExecution"
        },
        "startSubChain": {
          "$ref": "#/definitions/iotextypesStartSubChain",
          "title": "FedChain"
        },
        "stopSubChain": {
          "$ref": "#/definitions/iotextypesStopSubChain"
        },
        "putBlock": {
          "$ref": "#/definitions/iotextypesPutBlock"
        },
        "createDeposit": {
          "$ref": "#/definitions/iotextypesCreateDeposit"
        },
        "settleDeposit": {
          "$ref": "#/definitions/iotextypesSettleDeposit"
        },
        "createPlumChain": {
          "$ref": "#/definitions/iotextypesCreatePlumChain",
          "title": "PlumChain"
        },
        "terminatePlumChain": {
          "$ref": "#/definitions/iotextypesTerminatePlumChain"
        },
        "plumPutBlock": {
          "$ref": "#/definitions/iotextypesPlumPutBlock"
        },
        "plumCreateDeposit": {
          "$ref": "#/definitions/iotextypesPlumCreateDeposit"
        },
        "plumStartExit": {
          "$ref": "#/definitions/iotextypesPlumStartExit"
        },
        "plumChallengeExit": {
          "$ref": "#/definitions/iotextypesPlumChallengeExit"
        },
        "plumResponseChallengeExit": {
          "$ref": "#/definitions/iotextypesPlumResponseChallengeExit"
        },
        "plumFinalizeExit": {
          "$ref": "#/definitions/iotextypesPlumFinalizeExit"
        },
        "plumSettleDeposit": {
          "$ref": "#/definitions/iotextypesPlumSettleDeposit"
        },
        "plumTransfer": {
          "$ref": "#/definitions/iotextypesPlumTransfer"
        },
        "depositToRewardingFund": {
          "$ref": "#/definitions/iotextypesDepositToRewardingFund",
          "title": "Rewarding protocol actions"
        },
        "claimFromRewardingFund": {
          "$ref": "#/definitions/iotextypesClaimFromRewardingFund"
        },
        "grantReward": {
          "$ref": "#/definitions/iotextypesGrantReward"
        },
        "putPollResult": {
          "$ref": "#/definitions/iotextypesPutPollResult"
        }
      }
    },
...
    "iotextypesCandidate": {
      "type": "object",
      "properties": {
        "address": {
          "type": "string"
        },
        "votes": {
          "type": "string",
          "format": "byte"
        },
        "pubKey": {
          "type": "string",
          "format": "byte"
        },
        "rewardAddress": {
          "type": "string"
        }
      },
      "title": "Candidates and list of candidates"
    },
    "iotextypesCandidateList": {
      "type": "object",
      "properties": {
        "candidates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/iotextypesCandidate"
          }
        }
      }
    },
...
    "iotextypesClaimFromRewardingFund": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "iotextypesCreateDeposit": {
      "type": "object",
      "properties": {
        "chainID": {
          "type": "integer",
          "format": "int64"
        },
        "amount": {
          "type": "string"
        },
        "recipient": {
          "type": "string"
        }
      }
    },
    "iotextypesCreatePlumChain": {
      "type": "object",
      "title": "plum main chain APIs"
    },
    "iotextypesDepositToRewardingFund": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
...
    "iotextypesExecution": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string"
        },
        "contract": {
          "type": "string"
        },
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "iotextypesGrantReward": {
      "type": "object",
      "properties": {
        "type": {
          "$ref": "#/definitions/iotextypesRewardType"
        },
        "height": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "iotextypesMerkleRoot": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "iotextypesPlumChallengeExit": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "coinID": {
          "type": "string",
          "format": "uint64"
        },
        "challengeTransfer": {
          "type": "string",
          "format": "byte"
        },
        "challengeTransferBlockProof": {
          "type": "string",
          "format": "byte"
        },
        "challengeTransferBlockHeight": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesPlumCreateDeposit": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "amount": {
          "type": "string"
        },
        "recipient": {
          "type": "string"
        }
      }
    },
    "iotextypesPlumFinalizeExit": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "coinID": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesPlumPutBlock": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "height": {
          "type": "string",
          "format": "uint64"
        },
        "roots": {
          "type": "object",
          "additionalProperties": {
            "type": "string",
            "format": "byte"
          }
        }
      }
    },
    "iotextypesPlumResponseChallengeExit": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "coinID": {
          "type": "string",
          "format": "uint64"
        },
        "challengeTransfer": {
          "type": "string",
          "format": "byte"
        },
        "responseTransfer": {
          "type": "string",
          "format": "byte"
        },
        "responseTransferBlockProof": {
          "type": "string",
          "format": "byte"
        },
        "previousTransferBlockHeight": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesPlumSettleDeposit": {
      "type": "object",
      "properties": {
        "coinID": {
          "type": "string",
          "format": "uint64"
        }
      },
      "title": "plum sub chain APIs"
    },
    "iotextypesPlumStartExit": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "previousTransfer": {
          "type": "string",
          "format": "byte"
        },
        "previousTransferBlockProof": {
          "type": "string",
          "format": "byte"
        },
        "previousTransferBlockHeight": {
          "type": "string",
          "format": "uint64"
        },
        "exitTransfer": {
          "type": "string",
          "format": "byte"
        },
        "exitTransferBlockProof": {
          "type": "string",
          "format": "byte"
        },
        "exitTransferBlockHeight": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesPlumTransfer": {
      "type": "object",
      "properties": {
        "coinID": {
          "type": "string",
          "format": "uint64"
        },
        "denomination": {
          "type": "string",
          "format": "byte"
        },
        "owner": {
          "type": "string"
        },
        "recipient": {
          "type": "string"
        }
      }
    },
    "iotextypesPutBlock": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        },
        "height": {
          "type": "string",
          "format": "uint64"
        },
        "roots": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/iotextypesMerkleRoot"
          }
        }
      }
    },
    "iotextypesPutPollResult": {
      "type": "object",
      "properties": {
        "height": {
          "type": "string",
          "format": "uint64"
        },
        "candidates": {
          "$ref": "#/definitions/iotextypesCandidateList"
        }
      }
    },
...
    "iotextypesRewardType": {
      "type": "string",
      "enum": [
        "BlockReward",
        "EpochReward"
      ],
      "default": "BlockReward"
    },
...
    "iotextypesSettleDeposit": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string"
        },
        "recipient": {
          "type": "string"
        },
        "index": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesStartSubChain": {
      "type": "object",
      "properties": {
        "chainID": {
          "type": "integer",
          "format": "int64",
          "title": "TODO: chainID chould be assigned by system and returned via a receipt"
        },
        "securityDeposit": {
          "type": "string"
        },
        "operationDeposit": {
          "type": "string"
        },
        "startHeight": {
          "type": "string",
          "format": "uint64"
        },
        "parentHeightOffset": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "iotextypesStopSubChain": {
      "type": "object",
      "properties": {
        "chainID": {
          "type": "integer",
          "format": "int64"
        },
        "stopHeight": {
          "type": "string",
          "format": "uint64"
        },
        "subChainAddress": {
          "type": "string"
        }
      }
    },
    "iotextypesTerminatePlumChain": {
      "type": "object",
      "properties": {
        "subChainAddress": {
          "type": "string"
        }
      }
    },
    "iotextypesTransfer": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "title": "used by state-based model"
        },
        "recipient": {
          "type": "string"
        },
        "payload": {
          "type": "string",
          "format": "byte"
        }
      }
    },
...
  },
...
}